+++ packages/core/contracts/insured-bridge/BridgePool.sol
@@ -35,7 +35,7 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
     IERC20 public override l1Token;
 
     // Track the total number of relays and uniquely identifies relays.
-    uint256 public numberOfRelays;
+    uint32 public numberOfRelays;
 
     // Reserves that are unutilized and withdrawable.
     uint256 public liquidReserves;
@@ -75,12 +75,12 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
 
     // A Relay is linked to a L2 Deposit.
     struct RelayData {
-        uint256 relayId;
         RelayState relayState;
-        uint256 priceRequestTime;
-        uint64 realizedLpFeePct;
         address slowRelayer;
+        uint32 relayId;
+        uint64 realizedLpFeePct;
         address instantRelayer;
+        uint256 priceRequestTime;
     }
 
     // Associate deposits with pending relay data. When RelayState is Uninitialized, new relay attempts can be made for
@@ -96,31 +96,16 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
     event LiquidityAdded(address indexed token, uint256 amount, uint256 lpTokensMinted, address liquidityProvider);
     event LiquidityRemoved(address indexed token, uint256 amount, uint256 lpTokensBurnt, address liquidityProvider);
     event DepositRelayed(
-        uint256 indexed relayId,
-        uint8 chainId,
-        uint64 depositId,
-        address indexed l2Sender,
+        uint32 indexed relayId,
+        DepositData depositData,
         address slowRelayer,
-        address l1Recipient,
         address l1Token,
-        uint256 amount,
-        uint64 slowRelayFeePct,
-        uint64 instantRelayFeePct,
-        uint64 quoteTimestamp,
         uint64 realizedLpFeePct,
-        bytes32 indexed depositHash
+        bytes32 indexed depositHash,
+        bytes32 indexed relayHash
     );
     event RelaySpedUp(bytes32 indexed depositHash, address indexed instantRelayer);
-    event RelaySettled(
-        bytes32 indexed depositHash,
-        bytes32 indexed priceRequestAncillaryDataHash,
-        address indexed caller
-    );
-
-    modifier onlyFromOptimisticOracle() {
-        require(msg.sender == address(_getOptimisticOracle()), "Caller must be OptimisticOracle");
-        _;
-    }
+    event RelaySettled(bytes32 indexed depositHash, bytes32 indexed relayHash, address indexed caller);
 
     /**
      * @notice Construct the Bridge Pool
@@ -226,9 +211,7 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
     ) public nonReentrant() {
         // The realizedLPFeePct should never be greater than 0.5e18 and the slow and instant relay fees should never be
         // more than 0.25e18 each. Therefore, the sum of all fee types can never exceed 1e18 (or 100%).
-        require(slowRelayFeePct < 0.25e18);
-        require(instantRelayFeePct < 0.25e18);
-        require(realizedLpFeePct < 0.5e18);
+        require(slowRelayFeePct < 0.25e18 && instantRelayFeePct < 0.25e18 && realizedLpFeePct < 0.5e18);
 
         // Check if there is a pending relay for this deposit.
         DepositData memory depositData =
@@ -246,6 +229,8 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
 
         // If relay exists for deposit, check if it is disputed. If its disputed, then we can relay again, otherwise
         // the relay is pending valid and we cannot re-relay.
+        // Note: everything after the || gets called _only_ in the case that this relay comes after a previously
+        // disputed relay. Because of this, the getState call doesn't impact the gas usage in the happy path.
         require(
             relays[depositHash].relayState == RelayState.Uninitialized ||
                 relays[depositHash].relayState == RelayState.Disputed,
@@ -256,25 +241,26 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
         // instant relayer so that the l1Recipient cannot receive double payments. This means that once a relay is
         // disputed, it cant be sped up a second time (must finalize via the slow relay).
         uint256 priceRequestTime = getCurrentTime();
-        RelayData memory relayData =
-            RelayData({
-                relayId: numberOfRelays,
-                relayState: RelayState.Pending,
-                priceRequestTime: priceRequestTime,
-                realizedLpFeePct: realizedLpFeePct,
-                slowRelayer: msg.sender,
-                instantRelayer: relays[depositHash].instantRelayer
-            });
-        relays[depositHash] = relayData;
+
+        // Relay data is pulled out and set field-by-field because we're not setting _all_ fields.
+        RelayData storage relayData = relays[depositHash];
+
+        // This increments the storage variable at the same time as setting relayId.
+        uint32 relayId = numberOfRelays++;
+        relayData.relayId = relayId;
+        relayData.relayState = RelayState.Pending;
+        relayData.priceRequestTime = priceRequestTime;
+        relayData.realizedLpFeePct = realizedLpFeePct;
+        relayData.slowRelayer = msg.sender;
 
         bytes memory ancillaryData = getRelayAncillaryData(depositData, relayData);
         relayRequestAncillaryData[ancillaryData] = depositHash;
 
         // Sanity check that pool has enough balance to cover relay amount + proposer reward. Reward amount will be
         // paid on settlement after the OptimisticOracle price request has passed the challenge period.
+        uint256 proposerBond = _getProposerBond(amount);
         require(
-            l1Token.balanceOf(address(this)) >= amount + _getProposerBond(amount) &&
-                liquidReserves >= amount + _getProposerBond(amount),
+            l1Token.balanceOf(address(this)) >= amount + proposerBond && liquidReserves >= amount + proposerBond,
             "Insufficient pool balance"
         );
 
@@ -288,9 +274,15 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
         pendingReserves += amount; // Book off maximum liquidity used by this relay in the pending reserves.
 
         // We use an internal method to emit this event to overcome Solidity's "stack too deep" error.
-        _emitDepositRelayedEvent(depositData, realizedLpFeePct, depositHash);
-
-        numberOfRelays += 1;
+        emit DepositRelayed(
+            relayId,
+            depositData,
+            msg.sender,
+            address(l1Token),
+            realizedLpFeePct,
+            depositHash,
+            relayHash
+        );
     }
 
     /**
@@ -331,6 +323,7 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
 
         l1Token.safeTransferFrom(msg.sender, _depositData.l1Recipient, _depositData.amount - feesTotal);
 
+        // TODO: does this need more info?
         emit RelaySpedUp(depositHash, msg.sender);
     }
 
@@ -381,7 +374,11 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
         updateAccumulatedLpFees();
         allocateLpFees(_getAmountFromPct(relay.realizedLpFeePct, _depositData.amount));
 
-        emit RelaySettled(depositHash, keccak256(getRelayAncillaryData(_depositData, relay)), msg.sender);
+        emit RelaySettled(depositHash, relayHash, msg.sender);
+
+        delete relay.realizedLpFeePct;
+        delete relay.instantRelayer;
+        delete relay.priceRequestTime;
     }
 
     /**
@@ -458,7 +455,7 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
         updateAccumulatedLpFees(); // Accumulate all allocated fees from the last time this method was called.
         sync(); // Fetch any balance changes due to token bridging finalization and factor them in.
 
-        // ExchangeRate := (liquidReserves+utilizedReserves-undistributedLpFees)/lpTokenSupply
+        // ExchangeRate := (liquidReserves + utilizedReserves - undistributedLpFees) / lpTokenSupply
         // Note to accommodate negative utilizedReserves without using FixedPoint.Signed we need to do a bit of
         // branching logic. This is a gas optimization so we don't need to import this extra library logic.
         FixedPoint.Unsigned memory numerator =
@@ -534,72 +531,56 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
         view
         returns (bytes memory)
     {
-        // TODO: Consider adding BridgePool address to the ancillary data packet.
-        // TODO: Consider hashing all of the params that can be compared against the L2 Deposit contract into a single
-        // "relay ancillary data hash" to reduce storage costs.
-        bytes memory intermediateAncillaryData = "";
+        return
+            AncillaryData.appendKeyValueBytes32(
+                "",
+                "relayHash",
+                keccak256(
+                    abi.encode(
+                        _depositData.chainId,
+                        _depositData.depositId,
+                        _depositData.l1Recipient,
+                        _depositData.l2Sender,
+                        _depositData.amount,
+                        _depositData.slowRelayFeePct,
+                        _depositData.instantRelayFeePct,
+                        _depositData.quoteTimestamp,
+                        _relayData.relayId,
+                        _relayData.realizedLpFeePct,
+                        address(l1Token)
+                    )
+                )
+            );
+    }
 
-        // Add data inferred from the original deposit on L2:
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "chainId",
-            uint256(_depositData.chainId)
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "depositId",
-            uint256(_depositData.depositId)
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueAddress(
-            intermediateAncillaryData,
-            "l1Recipient",
-            _depositData.l1Recipient
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueAddress(
-            intermediateAncillaryData,
-            "l2Sender",
-            _depositData.l2Sender
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "amount",
-            _depositData.amount
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "slowRelayFeePct",
-            uint256(_depositData.slowRelayFeePct)
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "instantRelayFeePct",
-            uint256(_depositData.instantRelayFeePct)
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "quoteTimestamp",
-            uint256(_depositData.quoteTimestamp)
-        );
+    // Note: this method is identical to the one above, but it allows storage to be passed in, which saves some gas (3-4k)
+    // when called internally due to solidity not needing to copy the entire data structure and just lazily read data
+    // when requested.
+    function _getRelayAncillaryData(bytes32 relayHash) private pure returns (bytes memory) {
+        return AncillaryData.appendKeyValueBytes32("", "relayHash", relayHash);
+    }
 
-        // Add relay data.
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "relayId",
-            uint256(_relayData.relayId)
-        );
-        intermediateAncillaryData = AncillaryData.appendKeyValueUint(
-            intermediateAncillaryData,
-            "realizedLpFeePct",
-            uint256(_relayData.realizedLpFeePct)
-        );
-
-        // Add global state data stored by this contract:
-        intermediateAncillaryData = AncillaryData.appendKeyValueAddress(
-            intermediateAncillaryData,
-            "l1Token",
-            address(l1Token)
-        );
-        return intermediateAncillaryData;
+    function _getRelayHash(DepositData memory _depositData, RelayData storage _relayData)
+        private
+        view
+        returns (bytes32)
+    {
+        return
+            keccak256(
+                abi.encode(
+                    _depositData.chainId,
+                    _depositData.depositId,
+                    _depositData.l1Recipient,
+                    _depositData.l2Sender,
+                    _depositData.amount,
+                    _depositData.slowRelayFeePct,
+                    _depositData.instantRelayFeePct,
+                    _depositData.quoteTimestamp,
+                    _relayData.relayId,
+                    _relayData.realizedLpFeePct,
+                    address(l1Token)
+                )
+            );
     }
 
     /**************************************
@@ -705,30 +686,6 @@ contract BridgePool is Testable, BridgePoolInterface, ExpandedERC20, MultiCaller
             msg.sender,
             // Canonical value representing "True"; i.e. the proposed relay is valid.
             1e18
-        );
-    }
-
-    function _emitDepositRelayedEvent(
-        DepositData memory _depositData,
-        uint64 realizedLpFeePct,
-        bytes32 _depositHash
-    ) private {
-        // Emit only information that is not stored in this contract. The relay data associated with the `_depositHash`
-        // can be queried on-chain via the `relays` mapping keyed by `_depositHash`.
-        emit DepositRelayed(
-            numberOfRelays,
-            _depositData.chainId,
-            _depositData.depositId,
-            _depositData.l2Sender,
-            msg.sender,
-            _depositData.l1Recipient,
-            address(l1Token),
-            _depositData.amount,
-            _depositData.slowRelayFeePct,
-            _depositData.instantRelayFeePct,
-            _depositData.quoteTimestamp,
-            realizedLpFeePct,
-            _depositHash
         );
     }
 }