+++ packages/core/migrations/14_deploy_tokenfactory.js
@@ -1,8 +1,0 @@
-const TokenFactory = artifacts.require("TokenFactory");
-const { getKeysForNetwork, deploy } = require("./MigrationUtils");
-
-module.exports = async function (deployer, network, accounts) {
-  const keys = getKeysForNetwork(network, accounts);
-
-  await deploy(deployer, network, TokenFactory, { from: keys.deployer });
-};