+++ packages/core/migrations/1_initial_migration.js
@@ -1,7 +1,0 @@
-const Migrations = artifacts.require("./Migrations.sol");
-const { getKeysForNetwork, deploy } = require("./MigrationUtils");
-
-module.exports = async function (deployer, network, accounts) {
-  const keys = getKeysForNetwork(network, accounts);
-  await deploy(deployer, network, Migrations, { from: keys.deployer });
-};