+++ packages/merkle-distributor/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/UMAprotocol/protocol/compare/@uma/merkle-distributor@1.3.1...@uma/merkle-distributor@1.3.2) (2021-10-01)
+
+**Note:** Version bump only for package @uma/merkle-distributor
+
+## [1.3.1](https://github.com/UMAprotocol/protocol/compare/@uma/merkle-distributor@1.3.0...@uma/merkle-distributor@1.3.1) (2021-10-01)
+
+**Note:** Version bump only for package @uma/merkle-distributor
+
 # [1.3.0](https://github.com/UMAprotocol/protocol/compare/@uma/merkle-distributor@1.2.4...@uma/merkle-distributor@1.3.0) (2021-09-28)
 
 ### Bug Fixes