+++ packages/monitors/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.2](https://github.com/UMAprotocol/protocol/compare/@uma/monitors@2.6.1...@uma/monitors@2.6.2) (2021-10-01)
+
+**Note:** Version bump only for package @uma/monitors
+
+## [2.6.1](https://github.com/UMAprotocol/protocol/compare/@uma/monitors@2.6.0...@uma/monitors@2.6.1) (2021-10-01)
+
+**Note:** Version bump only for package @uma/monitors
+
 # [2.6.0](https://github.com/UMAprotocol/protocol/compare/@uma/monitors@2.5.1...@uma/monitors@2.6.0) (2021-09-28)
 
 ### Bug Fixes