+++ packages/optimistic-oracle/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.1](https://github.com/UMAprotocol/protocol/compare/@uma/optimistic-oracle@2.7.0...@uma/optimistic-oracle@2.7.1) (2021-10-01)
+
+**Note:** Version bump only for package @uma/optimistic-oracle
+
+# [2.7.0](https://github.com/UMAprotocol/protocol/compare/@uma/optimistic-oracle@2.6.0...@uma/optimistic-oracle@2.7.0) (2021-10-01)
+
+### Features
+
+- **optimistic-oracle-proposer:** Should not approve blacklisted identifiers ([#3415](https://github.com/UMAprotocol/protocol/issues/3415)) ([1027bb3](https://github.com/UMAprotocol/protocol/commit/1027bb3f89b0b0bc259ed119b951a66d1ec99fb1))
+
 # [2.6.0](https://github.com/UMAprotocol/protocol/compare/@uma/optimistic-oracle@2.5.1...@uma/optimistic-oracle@2.6.0) (2021-09-28)
 
 ### Features